Fork: Fix read fences with MailboxSynchronizer
The problem is when a texture (and the corresponding GLImage) is
deleted, the read fence is deleted and never clears.

In this workaround, we keep a "read fence for all deleted textures"
that is renewed every time a texture is deleted. This fence will have a
null GLImage client pointer, which will be cleared in the next
WillWrite.

BUG: 17690996
Change-Id: I067860a62a3e445ded6208906fa073f4f57239fe