Cherry pick: Cache viewport rect for tile priority in UpdateTiles
This is a cherry pick from chromium:
https://crrev.com/9178c77a0167ee7c73364b21a6f3a633029b7a44

Original CL description:
PictureLayerImpls in a tree should continue using the same viewport rect
for tile priority until UpdateTiles() is called with the new value of
viewport rect for tile priority. Otherwise, a PictureLayerImpl could be
in an inconsistent state that could lead to a wrong decision about
AllTilesRequiredForActivationAreReadyToDraw.

This CL caches the viewport rect for tile priority in PictureLayerImpl.

Conflicts:
  cc/layers/picture_layer_impl.h

BUG: 18915361
Change-Id: I7738cf793484fa46670388e0cf0e1d2cefbbddd4