Cherry pick: [WebView] Reword crash handler message.
This is a cherry pick from chromium:
https://crrev.com/86fca587a88a12c485c5b3bfaf7d3029561137af

Original CL description:
The current crash handler prints a message on logcat of the form
"### WebView crash. Version ..." whenever a crash is detected in an
Android app which happens to have a WebView.
Due to the in-process nature of the WebView, the crash handler is
triggered regardless of the actual cause of the crash, which might be
totally unrelated with the WebView or the chromium codebase at all.
The current wording is misleading as it erroneously suggests that the
crash is always due to the WebView.

BUG: 19293339
Change-Id: Ieeba42f19f7309fcbda5fa6ca0baa1cb33955092